requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
opencv-python-headless>=4.9.0
opencv-contrib-python-headless>=4.9.0
pyserial>=3.5
shapely>=2.0.0
numpy>=1.26.0
ultralytics>=8.1.0
pillow>=10.0.0
python-multipart>=0.0.9
